#68 - Moved the icons and rich text entry components
