+++ apps/spa-3ncount3r/src/components/elements/damage-condition-type/damage-condition-type.tsx
@@ -1,32 +1,32 @@
 import React from "react";
 import { Chip, SvgIcon } from "@mui/material";
 
-import { NecroticDamageIcon } from "../icons/damageType.necrotic.icon";
-import { ColdDamageIcon } from "../icons/damageType.cold.icon";
-import { LightningDamageIcon } from "../icons/damageType.lightning.icon";
-import { BludgeoningDamageIcon } from "../icons/damageType.bludgeoning.icon";
-import { PiercingDamageIcon } from "../icons/damageType.piercing.icon";
-import { SlashingDamageIcon } from "../icons/damageType.slashing.icon";
-import { MagicalBludgeoningDamageIcon } from "../icons/damageType.bludgeoning.magical.icon";
-import { MagicalPiercingDamageIcon } from "../icons/damageType.piercing.magical.icon";
-import { MagicalSlashingDamageIcon } from "../icons/damageType.slashing.magical.icon";
-import { PoisonDamageIcon } from "../icons/damageType.poison.icon";
-import { PsychicDamageIcon } from "../icons/damageType.psychic.icon";
-import { CharmedDamageIcon } from "../icons/damageType.charmed.icon";
-import { DeafenedDamageIcon } from "../icons/damageType.deafened.icon";
-import { ExhaustionDamageIcon } from "../icons/damageType.exhaustion.icon";
-import { FrightenedDamageIcon } from "../icons/damageType.frightened.icon";
-import { ParalyzedDamageIcon } from "../icons/damageType.paralyzed.icon";
-import { PetrifiedDamageIcon } from "../icons/damageType.petrified.icon";
-import { PoisonedDamageIcon } from "../icons/damageType.poisoned.icon";
-import { ProneDamageIcon } from "../icons/damageType.prone.icon";
-import { StunnedDamageIcon } from "../icons/damageType.stunned.icon";
-import { AcidDamageIcon } from "../icons/damageType.acid.icon";
-import { FireDamageIcon } from "../icons/damageType.fire.icon";
-import { RestrainedDamageIcon } from "../icons/damageType.restrained.icon";
-import { BlindedDamageIcon } from "../icons/damageType.blinded.icon";
+import { NecroticDamageIcon } from "../../icons/damageType.necrotic.icon";
+import { ColdDamageIcon } from "../../icons/damageType.cold.icon";
+import { LightningDamageIcon } from "../../icons/damageType.lightning.icon";
+import { BludgeoningDamageIcon } from "../../icons/damageType.bludgeoning.icon";
+import { PiercingDamageIcon } from "../../icons/damageType.piercing.icon";
+import { SlashingDamageIcon } from "../../icons/damageType.slashing.icon";
+import { MagicalBludgeoningDamageIcon } from "../../icons/damageType.bludgeoning.magical.icon";
+import { MagicalPiercingDamageIcon } from "../../icons/damageType.piercing.magical.icon";
+import { MagicalSlashingDamageIcon } from "../../icons/damageType.slashing.magical.icon";
+import { PoisonDamageIcon } from "../../icons/damageType.poison.icon";
+import { PsychicDamageIcon } from "../../icons/damageType.psychic.icon";
+import { CharmedDamageIcon } from "../../icons/damageType.charmed.icon";
+import { DeafenedDamageIcon } from "../../icons/damageType.deafened.icon";
+import { ExhaustionDamageIcon } from "../../icons/damageType.exhaustion.icon";
+import { FrightenedDamageIcon } from "../../icons/damageType.frightened.icon";
+import { ParalyzedDamageIcon } from "../../icons/damageType.paralyzed.icon";
+import { PetrifiedDamageIcon } from "../../icons/damageType.petrified.icon";
+import { PoisonedDamageIcon } from "../../icons/damageType.poisoned.icon";
+import { ProneDamageIcon } from "../../icons/damageType.prone.icon";
+import { StunnedDamageIcon } from "../../icons/damageType.stunned.icon";
+import { AcidDamageIcon } from "../../icons/damageType.acid.icon";
+import { FireDamageIcon } from "../../icons/damageType.fire.icon";
+import { RestrainedDamageIcon } from "../../icons/damageType.restrained.icon";
+import { BlindedDamageIcon } from "../../icons/damageType.blinded.icon";
 
-import { damageConditionChips } from "../../../../styles/details.styles";
+import { damageConditionChips } from "../../../styles/details.styles";
 
 interface DamageConditionTypeProps {
     key: string;
