+++ apps/spa-3ncount3r/src/components/icons/burrowingSpeed.icon.tsx
