+++ apps/spa-3ncount3r/src/components/icons/climbingSpeed.icon.tsx
