+++ apps/spa-3ncount3r/src/components/icons/damageType.fire.icon.tsx
