+++ apps/spa-3ncount3r/src/components/icons/damageType.paralyzed.icon.tsx
