+++ apps/spa-3ncount3r/src/components/icons/damageType.slashing.magical.icon.tsx
