+++ apps/spa-3ncount3r/src/components/icons/dice.d10.icon.tsx
