+++ apps/spa-3ncount3r/src/components/icons/dice.d20.icon.tsx
