+++ apps/spa-3ncount3r/src/components/icons/dice.d4.icon.tsx
