+++ apps/spa-3ncount3r/src/components/icons/dice.d6.icon.tsx
