+++ apps/spa-3ncount3r/src/components/icons/dice.d8.icon.tsx
