+++ apps/spa-3ncount3r/src/components/icons/flyingSpeed.icon.tsx
