+++ apps/spa-3ncount3r/src/components/icons/initiativeOrder.icon.tsx
