+++ apps/spa-3ncount3r/src/components/icons/movementSpeed.icon.tsx
