+++ apps/spa-3ncount3r/src/components/icons/swimmingSpeed.icon.tsx
