+++ apps/spa-3ncount3r/src/styles/details.styles.ts
